+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Add support for verifying with modulus/exponent components for RSA
 - Change API for both sign/verify to take a `Key` enum rather than bytes
 - Update to 2018 edition
+- Changed aud field type in Validation to `Option<HashSet<String>>`.  Audience 
+  validation now tests for "any-of-these" audience membership.
 
 ## 6.0.1 (2019-05-10)